+++ CHANGELOG.md
@@ -9,3 +9,5 @@
 - [emperror](https://github.com/goph/emperror) error handler constructor
 - [airbrake](https://github.com/airbrake/gobrake) compatible error handler constructor
 - Error handler stack for collecting multiple handlers under a single dependency
+- [opentracing](http://opentracing.io/) tracer constructor
+- [jaeger](https://github.com/jaegertracing) tracer constructor